Update módulo Soporte (#11)
* Update migraciones módulo Soporte

* Update modelos módulo Soporte

* Update seeders módulo Soporte

---------

+++ database/migrations/MER/2025_12_09_231636_create_estados_ticket_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('estados_ticket', function (Blueprint $table) {
-            $table->integer('cod', true);
+            $table->unsignedTinyInteger('cod', true);
             $table->string('des', 45)->nullable();
         });
     }
